+++ Установка/Quotes/Daily_quotes.json
@@ -1,0 +1,126 @@
+{
+	"1": {
+		"text": "Причина, по которой люди быстро сдаются - они смотрят сколько ещё нужно сделать вместо того, чтобы увидеть, как много уже сделано.",
+		"author": "Неизвестный"
+	},
+	"2": {
+		"text": "Спорт – это биение твоего сердца, твоё дыхание, это твой ритм жизни.",
+		"author": "Дейл Карнеги"
+	},
+	"3": {
+		"text": "Спортсмен может добиться высоких результатов и без природного таланта, но никогда без трудолюбия и упорства.",
+		"author": "Петр Квятковский"
+	},
+	"4": {
+		"text": "Спящему не достанется ничего, кроме снов.",
+		"author": "Серена Уильямс"
+	},
+	"5": {
+		"text": "Самое большое препятствие — страх. Самая большая ошибка — пасть духом.",
+		"author": "Джон Вуден"
+	},
+	"6": {
+		"text": "Многие люди никогда не сделают прорыв в своей жизни, потому что они отказались выйти из зоны комфорта и сделать шаг в неизвестность…",
+		"author": "Наполеон Хилл"
+	},
+	"7": {
+		"text": "Если ты не можешь, тогда должен. А если должен, значит сможешь.",
+		"author": "Тони Роббинс"
+	},
+	"8": {
+		"text": "Мечты не работают, пока не работаешь ты.",
+		"author": "Грейс Хоппер"
+	},
+	"9": {
+		"text": "Ты никогда не будешь неудачником, пока не перестанешь пытаться.",
+		"author": "Майк Дитка"
+	},
+	"10": {
+		"text": "То, что нас не убивает, делает нас сильнее.",
+		"author": "Фридрих Ницше"
+	},
+	"11": {
+		"text": "Все победы начинаются с победы над самим собой.",
+		"author": "Леонид Леонов"
+	},
+	"12": {
+		"text": "Если вы работаете над поставленными целями, то эти цели будут работать на вас.",
+		"author": "Д. Рон"
+	},
+	"13": {
+		"text": "Сильные падают — но поднимаются, проигрывают — но не сдаются, сражаются и побеждают. А слабые не способны даже поставить себе цель.",
+		"author": "Джон Вуден"
+	},
+	"14": {
+		"text": "Лучше маленькое дело, чем большой план без действия.",
+		"author": "Грейс Хоппер"
+	},
+	"15": {
+		"text": "Бедный, неудачный, несчастливый и нездоровый – это тот, кто часто использует слово «завтра».",
+		"author": "Роберт Кийосаки"
+	},
+	"16": {
+		"text": "Даже если вы падаете, вы все равно движетесь вперед.",
+		"author": "Виктор Киам"
+	},
+	"17": {
+		"text": "Ястребы взлетают гораздо выше, когда летят против ветра.",
+		"author": "Уинстон Черчилль"
+	},
+	"18": {
+		"text": "Дорога к успеху никогда не бывает окончательно построенной.",
+		"author": "Лили Томлин"
+	},
+	"19": {
+		"text": "Ничто не создаёт такого чувства собственного достоинства и уверенности в себе, как достижение.",
+		"author": "Томас Карлайл"
+	},
+	"20": {
+		"text": "Чтобы дойти до цели, надо идти.",
+		"author": "Оноре де Бальзак"
+	},
+	"21": {
+		"text": "Кто если не я? Когда если не сейчас?",
+		"author": "Арнольд Гелен"
+	},
+	"22": {
+		"text": "Пока у тебя есть попытка — ты не проиграл!",
+		"author": "Сергей Бубка"
+	},
+	"23": {
+		"text": "Постановка целей является первым шагом на пути превращения мечты в реальность.",
+		"author": "Тони Роббинс"
+	},
+	"24": {
+		"text": "Человеку всегда кажется, что в его силах намного меньше, чем он может на самом деле!",
+		"author": "Лес Браун"
+	},
+	"25": {
+		"text": "Кто хочет – ищет возможности, кто не хочет – ищет причины.",
+		"author": "Сократ"
+	},
+	"26": {
+		"text": "Дисциплина — это решение делать то, чего очень не хочется делать, чтобы достичь того, чего очень хочется достичь.",
+		"author": "Далай Лама"
+	},
+	"27": {
+		"text": "Сделай первый шаг, и ты поймешь, что не все так страшно.",
+		"author": "Сенека"
+	},
+	"28": {
+		"text": "Многое кажется невозможным, пока ты этого не сделаешь.",
+		"author": "Нельсон Мандела"
+	},
+	"29": {
+		"text": "Вам знакомо выражение «выше головы не прыгнешь»? Это заблуждение. Человек может все.",
+		"author": "Никола Тесла"
+	},
+	"30": {
+		"text": "Сделай 40 трудных минут, а не полтора часа чепухи.",
+		"author": "Джейсон Стейтем"
+	},
+	"31": {
+		"text": "Здоровый человек есть самое драгоценное произведение природы.",
+		"author": "Томас Карлейль"
+	}
+}
