+++ Установка/source/my_profile.json
@@ -1,0 +1,1 @@
+{"name": null, "gender": 0, "weight": null, "birthday": null, "photo_path": "./source/default_photo.jpg", "nice_days": 0, "goal": {"title": null, "type": null, "values": null}, "raiting_app": 0}
