+++ Установка/Прочитай меня.txt
@@ -1,0 +1,5 @@
+1. Выполните установку с помощью файла RWMInstallator
+2. В папку C:\Program Files (x86)\RunWithMe переместите папки new_design, Quotes, source
+3. Теперь можете запукать приложение
+
+Желаем успехов и новых рекордов!
